requirements.txt
fastapi
uvicorn
python-docx
sentence-transformers
faiss-cpu
numpy
openai
python-multipart
